Move package level function to object to fix api surface check
