+++ firebase-sessions/src/main/kotlin/com/google/firebase/sessions/ApplicationInfo.kt
@@ -16,7 +16,6 @@
 
 package com.google.firebase.sessions
 
-import com.google.firebase.FirebaseApp
 import com.google.firebase.encoders.json.NumberedEnum
 
 /** Enum denoting different development environments. */
@@ -58,17 +57,3 @@ internal data class ApplicationInfo(
   /** The android application information for the app. */
   val androidAppInfo: AndroidApplicationInfo,
 )
-
-internal fun getApplicationInfo(firebaseApp: FirebaseApp): ApplicationInfo {
-  val packageName = firebaseApp.applicationContext.packageName
-  val packageInfo = firebaseApp.applicationContext.packageManager.getPackageInfo(packageName, 0)
-
-  return ApplicationInfo(
-    appId = firebaseApp.options.applicationId,
-    deviceModel = "",
-    sessionSdkVersion = BuildConfig.VERSION_NAME,
-    logEnvironment = LogEnvironment.LOG_ENVIRONMENT_PROD,
-    androidAppInfo =
-      AndroidApplicationInfo(packageName = packageName, versionName = packageInfo.versionName)
-  )
-}
